Refactor implementation of top-level functions into `duet.api` module
This lets us keep the exported duet namespace clean without extraneous
"re-exports" of things like types.
